google_trends_pipeline/requirements.txt
pytrends
psycopg2-binary
python-dotenv
pytest
pylint